refactoring IoC

+++ src/main/java/es/upm/miw/doo/cohesion/Point.java
@@ -23,7 +23,7 @@ public class Point {
     }
 
     public double module() {
-        return Math.sqrt(this.x * this.x + this.y * this.y);
+        return Math.sqrt((double)this.x * this.x + this.y * this.y);
     }
 
     public double phase() {
